+++ PayPal/src/main/java/com/braintreepayments/api/PayPalClient.java
@@ -5,9 +5,9 @@ import android.content.Intent;
 import android.net.Uri;
 import android.text.TextUtils;
 
+import androidx.activity.ComponentActivity;
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
-import androidx.fragment.app.FragmentActivity;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -56,7 +56,7 @@ public class PayPalClient {
     /**
      * Starts the PayPal payment flow by creating a {@link PayPalPaymentAuthRequestParams} to be
      * used to launch the PayPal web authentication flow in
-     * {@link PayPalLauncher#launch(FragmentActivity, PayPalPaymentAuthRequestParams)}.
+     * {@link PayPalLauncher#launch(ComponentActivity, PayPalPaymentAuthRequest.ReadyToLaunch)}.
      *
      * @param context       Android Context
      * @param payPalRequest a {@link PayPalRequest} used to customize the request.
@@ -176,12 +176,12 @@ public class PayPalClient {
 
     /**
      * After receiving a result from the PayPal web authentication flow via
-     * {@link PayPalLauncher#handleReturnToAppFromBrowser(Context, Intent)}, pass the
+     * {@link PayPalLauncher#handleReturnToAppFromBrowser(PayPalPendingRequest.Started, Intent)}, pass the
      * {@link PayPalPaymentAuthResult} returned to this method to tokenize the PayPal account and
      * receive a {@link PayPalAccountNonce} on success.
      *
      * @param paymentAuthResult a {@link PayPalPaymentAuthResult} received in the callback
-     *                                  of {@link PayPalLauncher#PayPalLauncher(PayPalLauncherCallback)}
+     *                                  of {@link PayPalLauncher#PayPalLauncher()}
      * @param callback                  {@link PayPalTokenizeCallback}
      */
     public void tokenize(@NonNull PayPalPaymentAuthResult paymentAuthResult,